Add TaskScope and let TaskGroup subclass it
The rewritten TaskGroup still passes all existing test cases.
